JVM IR: optimize special method / signature computation in BridgeLowering
First of all, put caches of BridgeLowering into JvmBackendContext. Since
method hierarchies can span several files, and a new instance of
BridgeLowering is created to lower each file, keeping the cached special
bridges and method signatures for the whole module makes BridgeLowering
faster. With this change, BridgeLowering takes 2.0% instead of 3.8% of
all compiler time on a sample project.

Also, do not attempt to compute special bridges if the method is
irrelevant, which can be deduced by its name. This brings BridgeLowering
time to 1.5% of all compiler time.